requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0.post1
pydantic>=2.4.2
pydantic-settings>=2.0.3
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.1
google-generativeai>=0.7.2
redis>=5.0.1
celery>=5.3.6
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1
python-dateutil>=2.8.2
tenacity>=8.2.3
python-json-logger>=2.0.7
pytest-mock>=3.12.0
